Fixed bug with local variable initialization. All tests pass
In the previous version local variables were not properly initialized to 0. This error is fixed now.
Note that programs that don't include a Sys.vm file in the directory (and have R0, etc. declared in the tst file) require that line 37 in Main.java be commented out to prevent bootstrap code from being added to the asm file.