Handles input directories with/without Sys.vm better
If a directory is provided, and this directory contains a Sys.vm file, then bootstrap code is added to the assembly output. If a directory does not contain Sys.vm that usually means that the user set the required addresses in the test file, so no bootstrap code will be added to the output file.